All Leaks issues fixed, commit 18 JUL, 14:29

+++ inc/libft/src/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: gfrancis <gfrancis@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/13 15:16:05 by gfrancis          #+#    #+#             */
-/*   Updated: 2023/07/11 12:39:01 by gfrancis         ###   ########.fr       */
+/*   Updated: 2023/07/18 10:19:31 by gfrancis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	if (!str)
